Konsoliderer operasjoner fra overflødige StaticDataControllerV2 inn i StaticDataControllerV1, etter pattern fra OrkestreringsController.
